Add files via upload
In this module, we are adding new user.  For that we have to add his 64 images for further model creation of face recognition. This images are taken from camera, user has to give extra expression. model will automatically add that images to database and model for face recognition will created.